refactor : Observability
- Renamed logger_adaptors.py to logger_adaptor.py
- Moved observability and related adaptors to a seperate module inside /application-sdk
