+++ application_sdk/clients/temporal.py
@@ -19,7 +19,6 @@ from temporalio.worker.workflow_sandbox import (
 )
 
 from application_sdk.clients.workflow import WorkflowClient
-from application_sdk.common.logger_adaptors import get_logger
 from application_sdk.constants import (
     APPLICATION_NAME,
     MAX_CONCURRENT_ACTIVITIES,
@@ -28,6 +27,7 @@ from application_sdk.constants import (
     WORKFLOW_NAMESPACE,
     WORKFLOW_PORT,
 )
+from application_sdk.observability.logger_adaptor import get_logger
 from application_sdk.outputs.eventstore import (
     ActivityEndEvent,
     ActivityStartEvent,
