+++ application_sdk/docgen/__init__.py
@@ -3,11 +3,11 @@ import os
 from datetime import datetime
 from typing import List
 
-from application_sdk.common.logger_adaptors import get_logger
 from application_sdk.docgen.exporters.mkdocs import MkDocsExporter
 from application_sdk.docgen.models.export.page import Page
 from application_sdk.docgen.parsers.directory import DirectoryParser
 from application_sdk.docgen.parsers.manifest import ManifestParser
+from application_sdk.observability.logger_adaptor import get_logger
 
 logger = get_logger(__name__)
