+++ application_sdk/inputs/iceberg.py
@@ -2,8 +2,8 @@ from typing import TYPE_CHECKING, AsyncIterator, Iterator, Optional, Union
 
 from pyiceberg.table import Table
 
-from application_sdk.common.logger_adaptors import get_logger
 from application_sdk.inputs import Input
+from application_sdk.observability.logger_adaptor import get_logger
 
 logger = get_logger(__name__)
