+++ application_sdk/inputs/sql_query.py
@@ -2,8 +2,8 @@ import asyncio
 import concurrent
 from typing import TYPE_CHECKING, AsyncIterator, Iterator, Optional, Union
 
-from application_sdk.common.logger_adaptors import get_logger
 from application_sdk.inputs import Input
+from application_sdk.observability.logger_adaptor import get_logger
 
 logger = get_logger(__name__)
