+++ application_sdk/observability/metrics_adaptor.py
@@ -11,8 +11,8 @@ from opentelemetry.sdk.metrics.export import PeriodicExportingMetricReader
 from opentelemetry.sdk.resources import Resource
 from pydantic import BaseModel
 
-from application_sdk.common.logger_adaptors import get_logger
-from application_sdk.common.observability import AtlanObservability
+from application_sdk.observability.logger_adaptor import get_logger
+from application_sdk.observability.observability import AtlanObservability
 from application_sdk.constants import (
     ENABLE_OTLP_METRICS,
     METRICS_BATCH_SIZE,
