+++ application_sdk/transformers/atlas/sql.py
@@ -11,7 +11,7 @@ from pyatlan.model import assets
 from pyatlan.model.enums import AtlanConnectorType
 from pyatlan.utils import init_guid, validate_required_fields
 
-from application_sdk.common.logger_adaptors import get_logger
+from application_sdk.observability.logger_adaptor import get_logger
 from application_sdk.transformers.common.utils import build_atlas_qualified_name
 
 logger = get_logger(__name__)
