+++ application_sdk/transformers/query/__init__.py
@@ -6,7 +6,7 @@ import daft
 import yaml
 from pyatlan.model.enums import AtlanConnectorType
 
-from application_sdk.common.logger_adaptors import get_logger
+from application_sdk.observability.logger_adaptor import get_logger
 from application_sdk.transformers import TransformerInterface
 from application_sdk.transformers.common.utils import (
     get_yaml_query_template_path_mappings,
