+++ application_sdk/workflows/query_extraction/sql.py
@@ -13,9 +13,9 @@ from temporalio.common import RetryPolicy
 
 from application_sdk.activities.query_extraction.sql import SQLQueryExtractionActivities
 from application_sdk.clients.sql import BaseSQLClient
-from application_sdk.common.logger_adaptors import get_logger
 from application_sdk.constants import APPLICATION_NAME
 from application_sdk.inputs.statestore import StateStoreInput
+from application_sdk.observability.logger_adaptor import get_logger
 from application_sdk.workflows.query_extraction import QueryExtractionWorkflow
 
 logger = get_logger(__name__)
