+++ examples/application_sql_with_custom_pyatlan_transformer.py
@@ -38,8 +38,8 @@ from application_sdk.application.metadata_extraction.sql import (
     BaseSQLMetadataExtractionApplication,
 )
 from application_sdk.clients.sql import BaseSQLClient
-from application_sdk.common.logger_adaptors import get_logger
 from application_sdk.handlers.sql import BaseSQLHandler
+from application_sdk.observability.logger_adaptor import get_logger
 from application_sdk.transformers.atlas import AtlasTransformer
 from application_sdk.transformers.atlas.sql import Column, Procedure, Table
 from application_sdk.workflows.metadata_extraction.sql import (
