+++ examples/application_subscriber.py
@@ -7,9 +7,9 @@ from temporalio import activity, workflow
 from application_sdk.activities import ActivitiesInterface
 from application_sdk.activities.common.utils import auto_heartbeater
 from application_sdk.clients.utils import get_workflow_client
-from application_sdk.common.logger_adaptors import get_logger
 from application_sdk.constants import APPLICATION_NAME
 from application_sdk.inputs.statestore import StateStoreInput
+from application_sdk.observability.logger_adaptor import get_logger
 from application_sdk.outputs.eventstore import (
     WORKFLOW_END_EVENT,
     AtlanEvent,
